separates shared memory manager and shared memory implementation into separate files.
DlgSharedMemoryManager now handles duplicate and unregistered session requests.
